dailyfresh/requirements.txt
amqp==1.4.9
anyjson==0.3.3
backcall==0.1.0
billiard==3.3.0.23
celery==3.1.25
certifi==2018.11.29
chardet==3.0.4
decorator==4.3.2
Django==2.1.7
django-celery==3.1.17
django-haystack==2.8.1
django-redis==4.10.0
django-redis-sessions==0.5.6
django-tinymce==3.0.2
fdfs-client-py==1.2.6
idna==2.8
ipython==7.3.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.13.3
jieba==0.39
kombu==3.0.37
mutagen==1.42.0
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
Pillow==3.4.1
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pycryptodomex==3.7.2
Pygments==2.3.1
PyMySQL==0.9.3
python-alipay-sdk==1.10.0
pytz==2018.9
redis==2.10.6
redis-py-cluster==1.3.6
requests==2.21.0
six==1.12.0
traitlets==4.3.2
urllib3==1.24.1
uWSGI==2.0.18
vine==1.2.0
wcwidth==0.1.7
Whoosh==2.7.4
